Separate out script to create tables Add relative_path column to dataset table Start adding things needed for a proper package
