+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup(setup_requires=['setuptools_scm'])
